Install pybind11 to CI & Fix cibuildwheel package directory

+++ bindings/python/tools/prepare_build_environment.sh
@@ -8,6 +8,16 @@ ICU_VERSION=${ICU_VERSION:-64.2}
 SENTENCEPIECE_VERSION=${SENTENCEPIECE_VERSION:-0.1.8}
 PYBIND11_VERSION=${PYBIND11_VERSION:-2.4.3}
 
+# Install pybind11
+pip install pybind11==$PYBIND11_VERSION
+
+# Install CMake
+pip install "cmake==3.13.*"
+
+# Skip double installation of libraries
+if [ -f built-$PYTHON_ARCH ]; then exit 0; fi
+touch built-$PYTHON_ARCH
+
 # Install ICU
 curl -L -O https://github.com/unicode-org/icu/releases/download/release-${ICU_VERSION/./-}/icu4c-${ICU_VERSION/./_}-src.tgz
 tar xf icu4c-*-src.tgz
@@ -15,9 +25,6 @@ cd icu/source
 CFLAGS="-fPIC" CXXFLAGS="-fPIC" ./configure --disable-shared --enable-static
 make -j2 install
 cd $ROOT_DIR
-
-# Install CMake
-pip install "cmake==3.13.*"
 
 # Build SentencePiece
 curl -L -o sentencepiece-${SENTENCEPIECE_VERSION}.tar.gz -O https://github.com/google/sentencepiece/archive/v${SENTENCEPIECE_VERSION}.tar.gz
@@ -33,6 +40,3 @@ mkdir build; cd build
 cmake -DCMAKE_BUILD_TYPE=Release -DLIB_ONLY=ON -DWITH_ICU=ON ..
 make -j2 install
 cd $ROOT_DIR
-
-# Install pybind11
-pip install pybind11==$PYBIND11_VERSION
